% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rie.R
\name{rie}
\alias{rie}
\alias{rie_vec}
\title{Implementation of the Robust Initial Enhancement score.}
\usage{
rie(
  data,
  truth,
  ...,
  alpha = 20,
  estimator = NULL,
  na_rm = TRUE,
  event_level = "first"
)

rie(
  data,
  truth,
  ...,
  alpha = 20,
  estimator = NULL,
  na_rm = TRUE,
  event_level = "first"
)

rie_vec(
  truth,
  estimate,
  alpha = 20,
  estimator = NULL,
  event_level = "first",
  na_rm = TRUE
)
}
\arguments{
\item{data}{A data.frame containing the truth and estimate columns.}

\item{truth}{The column identifier for the true class results (that is a factor). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a factor vector.}

\item{...}{A set of unquoted column names or one or more dplyr selector functions to choose which variables contain the class probabilities. If truth is binary, only 1 column should be selected. Otherwise, there should be as many columns as factor levels of truth.}

\item{alpha}{The early recognition parameter (defaults to 20 or as in enrichvs::rie())}

\item{estimator}{One of "binary", "hand_till", "macro", or "macro_weighted" to specify the type of averaging to be done. "binary" is only relevant for the two class case. The others are general methods for calculating multiclass metrics. The default will automatically choose "binary" or "hand_till" based on truth.}

\item{na_rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{event_level}{A single string. Either "first" or "second" to specify which level of truth to consider as the "event". This argument is only applicable when estimator = "binary". The default uses an internal helper that generally defaults to "first", however, if the deprecated global option yardstick.event_first is set, that will be used instead with a warning.}
}
\value{
A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For rie_vec(), a single numeric value (or NA).
}
\description{
Calculates a metric often used in virtual screening - Robust Initial Enhancement (RIE)
}
\references{
Sheridan RP, Singh SB, Fluder EM, Kearsley SK. Protocols for bridging the peptide to nonpeptide gap in topological similarity searches. J Chem Inf Comput Sci. 2001 Sep;41(5):1395–406.
}
\seealso{
\code{\link[=ef]{ef()}} for computing the enrichment factor and \code{\link[=bedroc]{bedroc()}} for computing the Boltzmann-enhanced discrimination ROC.
}
