% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truth_event_level}
\alias{truth_event_level}
\title{Get the correct event level}
\usage{
truth_event_level(truth, event_level)
}
\arguments{
\item{truth}{A factor of the truth data}

\item{event_level}{The level of the factor that contains the event (the class of interest)}
}
\value{
String
}
\description{
Get the correct event level
}
\examples{
x <- factor(c('TRUE', 'FALSE', 'TRUE', 'TRUE'))
truth_event_level(x, 'first')
}
